Move tree utilities into main utils module
