+++ src/esquilax/utils/__init__.py
@@ -2,6 +2,6 @@
 Modelling and algorithm utility functions
 """
 
-from . import functions, graph, space
+from . import functions, graph, space, tree
 from .graph import sort_edges
 from .space import shortest_distance, shortest_vector
