+++ tests/test_ml/test_ml_utils.py
@@ -4,7 +4,6 @@ import jax.numpy as jnp
 import pytest
 
 from esquilax import ml
-from esquilax.ml import common
 
 
 @pytest.fixture
@@ -61,22 +60,3 @@ def test_sample_action_actions(broadcast, random_apply_fun):
         action = ml.sample_actions(random_apply_fun, False, k, params, obs)
 
     assert action.shape == (n_agents,)
-
-
-def test_tree_key_split():
-    k = jax.random.PRNGKey(451)
-
-    a = {"a": 1, "b": (2, 3)}
-    b = common.key_tree_split(k, a)
-
-    assert jax.tree.structure(a) == jax.tree.structure(b)
-    assert not jnp.array_equal(b["a"], b["b"][0])
-    assert not jnp.array_equal(b["a"], b["b"][1])
-
-
-def test_tuple_tree_transpose():
-    a = {"a": 1, "b": 2}
-    b = {"a": (1, (2, 3)), "b": (4, (5, 6))}
-    c = common.transpose_tree_of_tuples(a, b, 2)
-
-    assert c == ({"a": 1, "b": 4}, {"a": (2, 3), "b": (5, 6)})
