+++ tests/test_utils.py
@@ -1,4 +1,5 @@
 import chex
+import jax
 import jax.numpy as jnp
 import pytest
 
@@ -150,3 +151,22 @@ def test_keyword_arguments():
         pass
 
     assert utils.functions.get_keyword_args(bar) == []
+
+
+def test_tree_key_split():
+    k = jax.random.PRNGKey(451)
+
+    a = {"a": 1, "b": (2, 3)}
+    b = utils.tree.key_tree_split(k, a)
+
+    assert jax.tree.structure(a) == jax.tree.structure(b)
+    assert not jnp.array_equal(b["a"], b["b"][0])
+    assert not jnp.array_equal(b["a"], b["b"][1])
+
+
+def test_tuple_tree_transpose():
+    a = {"a": 1, "b": 2}
+    b = {"a": (1, (2, 3)), "b": (4, (5, 6))}
+    c = utils.tree.transpose_tree_of_tuples(a, b, 2)
+
+    assert c == ({"a": 1, "b": 4}, {"a": (2, 3), "b": (5, 6)})
